Trait zk_pass::chaum_pedersen::ChaumPedersen

pub trait ChaumPedersen {
    type Secret;
    type Response;
    type Challenge;
    type GroupParameters;
    type CommitParameters;
    type CommitmentRandom;

    // Required methods
    fn calculate_commitment(
        params: &Self::GroupParameters,
        x: &Self::Secret
    ) -> (Self::CommitParameters, Self::CommitmentRandom)
       where Self: Sized;
    fn challenge(params: &Self::GroupParameters) -> Self::Challenge
       where Self: Sized;
    fn calculate_response(
        params: &Self::GroupParameters,
        k: &Self::CommitmentRandom,
        c: &Self::Challenge,
        x: &Self::Secret
    ) -> Self::Response
       where Self: Sized;
    fn verify(
        params: &Self::GroupParameters,
        s: &Self::Response,
        c: &Self::Challenge,
        cp: &Self::CommitParameters
    ) -> bool
       where Self: Sized;
}

A trait defining the interface for the Chaum-Pedersen zero-knowledge protocol.

This trait provides the necessary methods for implementing the Chaum-Pedersen protocol, which is a cryptographic protocol for proving knowledge of a secret without revealing it.

Required Associated Types

type Secret

The type representing the secret to be proven.

type Response

The type representing the response in the protocol.

type Challenge

The type representing the challenge in the protocol.

type GroupParameters

The type representing the group parameters used in the protocol.

type CommitParameters

The type representing the commitment parameters in the protocol.

type CommitmentRandom

The type representing the commitment randomness in the protocol.

Required Methods

fn calculate_commitment( params: &Self::GroupParameters, x: &Self::Secret ) -> (Self::CommitParameters, Self::CommitmentRandom)where Self: Sized,

Calculates the commitment in the Chaum-Pedersen protocol.

Arguments

params - Group parameters used in the protocol.

x - The secret value for which the commitment is calculated.

Returns

A tuple containing the commitment parameters and the commitment randomness.

fn challenge(params: &Self::GroupParameters) -> Self::Challengewhere Self: Sized,

Generates a challenge in the Chaum-Pedersen protocol.

Arguments

params - Group parameters used in the protocol.

Returns

The challenge value used in the protocol.

fn calculate_response( params: &Self::GroupParameters, k: &Self::CommitmentRandom, c: &Self::Challenge, x: &Self::Secret ) -> Self::Responsewhere Self: Sized,

Calculates the response in the Chaum-Pedersen protocol.

Arguments

params - Group parameters used in the protocol.

k - The commitment randomness used in the protocol.

c - The challenge value used in the protocol.

x - The secret value for which the response is calculated.

Returns

The response value in the protocol.

fn verify( params: &Self::GroupParameters, s: &Self::Response, c: &Self::Challenge, cp: &Self::CommitParameters ) -> boolwhere Self: Sized,

Verifies the response in the Chaum-Pedersen protocol.

Arguments

params - Group parameters used in the protocol.

s - The response value to be verified.

c - The challenge value used in the protocol.

cp - The commitment parameters used in the protocol.

Returns

A boolean indicating whether the verification was successful.