Trait zk_pass::rand::RandomGenerator

pub trait RandomGenerator<T> {
    // Required method
    fn generate_random() -> Result<T, Box<dyn Error>>;
}

Defines a trait for generating random values of a given type.

This trait is intended to abstract the generation of random values for different types, providing a uniform interface.

Required Methods

fn generate_random() -> Result<T, Box<dyn Error>>

Generates a random value of type T.

Returns

A Result containing the random value of type T, or an error if the generation fails.

Errors

Returns an error if the random value generation fails.